refactor(deploy): extract poll core into status.ts
