+++ packages/cli-core/src/commands/deploy/index.ts
@@ -1,11 +1,9 @@
 import { isAgent } from "../../mode.ts";
 import { isInsideGutter, log } from "../../lib/log.ts";
-import { sleep } from "../../lib/sleep.ts";
-import { bar, intro, outro, withSpinner, type SpinnerControls } from "../../lib/spinner.ts";
+import { bar, intro, outro, withSpinner } from "../../lib/spinner.ts";
 import {
   CliError,
   ERROR_CODE,
-  PlapiError,
   UserAbortError,
   isPromptExitError,
   throwUsageError,
@@ -19,7 +17,6 @@ import {
   getApplicationDomainStatus,
   listApplicationDomains,
   patchInstanceConfig,
-  triggerApplicationDomainDNSCheck,
   type ApplicationDomain,
   type CnameTarget,
   type DomainStatusResponse,
@@ -28,12 +25,9 @@ import {
 import {
   INTRO_PREAMBLE,
   OAUTH_SECTION_INTRO,
-  type DeployComponentStatus,
   type DeployPlanStep,
-  DEPLOY_COMPONENT_ORDER,
   deployComponentLabels,
   deployComponentStatus,
-  deployStatusRetryMessage,
   deployStatusPendingFooter,
   domainAssociationSummary,
   bindZoneFile,
@@ -72,10 +66,7 @@ import {
   type DeployContext,
   type DeployOperationState,
 } from "./state.ts";
-
-const DEPLOY_STATUS_INITIAL_RETRY_DELAY_MS = 3000;
-const DEPLOY_STATUS_MAX_RETRIES = 5;
-const DEPLOY_STATUS_BACKOFF_FACTOR = 2;
+import { waitForDeployStatus, type DeployStatusOutcome } from "./status.ts";
 
 type DeployOptions = Record<string, never>;
 
@@ -639,94 +630,15 @@ async function runDnsVerification(
   }
 }
 
-type DeployStatusOutcome =
-  | { verified: true; status: DeployComponentStatus }
-  | { verified: false; status: DeployComponentStatus };
-
 async function pollDeployStatus(
   appId: string,
   domainIdOrName: string,
   domain: string,
 ): Promise<DeployStatusOutcome> {
-  await triggerDeployStatusCheck(appId, domainIdOrName);
-  let response = await mapDeployError(getApplicationDomainStatus(appId, domainIdOrName));
-  let status = deployComponentStatusFromDomainStatus(response);
-  for (const component of DEPLOY_COMPONENT_ORDER) {
-    let retriesRemaining = DEPLOY_STATUS_MAX_RETRIES;
-    let nextRetryDelay = DEPLOY_STATUS_INITIAL_RETRY_DELAY_MS;
-    const labels = deployComponentLabels(component, domain);
-    const flipped = await withSpinner(labels.progress, async (spinner) => {
-      if (status[component]) return true;
-      while (retriesRemaining > 0) {
-        await sleepWithRetryCountdown(
-          labels.progress,
-          DEPLOY_STATUS_MAX_RETRIES - retriesRemaining + 1,
-          DEPLOY_STATUS_MAX_RETRIES,
-          nextRetryDelay,
-          spinner,
-        );
-        retriesRemaining--;
-        nextRetryDelay *= DEPLOY_STATUS_BACKOFF_FACTOR;
-        response = await mapDeployError(getApplicationDomainStatus(appId, domainIdOrName));
-        status = deployComponentStatusFromDomainStatus(response);
-        if (status[component]) return true;
-      }
-      return false;
-    });
-    if (!flipped) return { verified: false, status };
-    log.success(labels.done);
-  }
-
-  if (response.status !== "complete") {
-    return { verified: false, status };
-  }
-  return { verified: true, status };
-}
-
-async function sleepWithRetryCountdown(
-  message: string,
-  currentRetry: number,
-  totalRetries: number,
-  delayMs: number,
-  spinner: SpinnerControls,
-): Promise<void> {
-  let remainingMs = delayMs;
-  while (remainingMs > 0) {
-    const tickMs = Math.min(1000, remainingMs);
-    spinner.update(
-      deployStatusRetryMessage(message, currentRetry, totalRetries, Math.ceil(remainingMs / 1000)),
-    );
-    await sleep(tickMs);
-    remainingMs -= tickMs;
-  }
-}
-
-async function triggerDeployStatusCheck(appId: string, domainIdOrName: string): Promise<void> {
-  try {
-    await mapDeployError(triggerApplicationDomainDNSCheck(appId, domainIdOrName));
-  } catch (error) {
-    if (error instanceof PlapiError && error.status === 409 && error.code === "conflict") {
-      log.debug("DNS check is already in flight; continuing to poll domain status.");
-      return;
-    }
-    throw error;
-  }
-}
-
-function deployComponentStatusFromDomainStatus(
-  response: DomainStatusResponse,
-): DeployComponentStatus {
-  return {
-    dns: checkStatusComplete(response.dns),
-    ssl: checkStatusComplete(response.ssl),
-    mail: checkStatusComplete(response.mail),
-  };
-}
-
-function checkStatusComplete(check: { status: string; required?: boolean } | undefined): boolean {
-  if (!check) return false;
-  if (check.required === false) return true;
-  return check.status === "complete";
+  return waitForDeployStatus(appId, domainIdOrName, domain, {
+    runComponent: (_component, progressLabel, work) => withSpinner(progressLabel, work),
+    onComponentDone: (component) => log.success(deployComponentLabels(component, domain).done),
+  });
 }
 
 async function offerBindZoneExport(
